Allow `with`'s `to_xml` to fail.
This should be a rare case, but in case some values can't be written out it would be nice to have the option to return an error rather than having to panic. It is a fairly low cost that other users need to wrap their `to_xml` adaptor with `Ok`.